+++ src/modules/home/HomeView.js
@@ -8,6 +8,7 @@ import {
   View,
 } from 'react-native';
 
+import ProgressBar from 'react-native-progress/Bar';
 import Icon from 'react-native-vector-icons/MaterialIcons';
 
 import SectionList from './SectionList';
@@ -78,8 +79,8 @@ export default class HomeView extends Component {
     }
 
     return (
-      <View>
-        <Text> loading... </Text>
+      <View style={styles.loading}>
+        <ProgressBar progress={0.3} width={200} />
       </View>
     );
   }